gosper {hypergeo}	R Documentation

Evaluation of the hypergeometric function using Gosper's method

Description

Evaluation of the hypergeometric function using Gosper's method

Usage

hypergeo_gosper(A, B, C, z, tol = 0, maxiter = 2000)

Arguments

A,B,C
Parameters (real or complex)

z
Complex argument

tol
tolerance (passed to GCF())

maxiter
maximum number of iterations

Details

Gosper provides a three-term recurrence which converges when z is close to a critical point.

Bill Gosper asserts that the recursion holds for values of z which are inside the cardioid (sqrt(8)*cos(t)-2*cos(2t), sqrt(8)*sin(t)-2*sin(2t)) (see examples section).

It is suggested that the recursion should only be used when the auxiliary parameters A, B,C are all <=12 in absolute value.

Author(s)

R code by Robin K. S. Hankin, transcribed from maxima code posted by Richard Fateman, who credited Bill Gosper

References

Original email was archived at https://www.ma.utexas.edu/pipermail/maxima/2006/000126.html but does not appear there now; and the wayback machine doesn't find it either.

See Also

hypergeo_contfrac

Examples

hypergeo_gosper(1.1,5.1,3.1,crit())

# Compare MMA: -0.192225 + 0.692328 I

t <- seq(from=0,to=2i*pi,len=100)
plot(exp(t)*(sqrt(8)-exp(t)),asp=1,type='l')
points(crit())

[Package hypergeo version 1.2-14 Index]