hypergeo-package {hypergeo}	R Documentation

The hypergeometric function

Description

The hypergeometric function for the whole complex plane

Details

Package: 	 hypergeo

Type: 	 Package

Version: 	 1.0

Date: 	 2008-04-16

License: 	 GPL

The front end function of the package is hypergeo(): depending on the value of the parameters, this executes one or more of many sub-functions.

Author(s)

Robin K. S. Hankin

Maintainer: <hankin.robin@gmail.com>

References

M. Abramowitz and I. A. Stegun 1965. Handbook of mathematical functions. New York: Dover

Examples

hypergeo(1.1,2.3,1.9 , 1+6i)

options(showHGcalls = TRUE)    # any non-null value counts as TRUE
hypergeo(4.1, 3.1, 5.1, 1+1i)  # shows trace back
options(showHGcalls = FALSE)   # reset

[Package hypergeo version 1.2-14 Index]