hypergeo_A_nonpos_int {hypergeo}	R Documentation

Hypergeometric functions for integer arguments

Description

Hypergeometric functions for A and/or B being integers

Usage

hypergeo_A_nonpos_int(A, B, C, z, tol = 0)
hypergeo_AorB_nonpos_int(A, B, C, z, tol = 0)

Arguments

A,B,C
Parameters for the hypergeometric function

z
Primary complex argument

tol
tolerance

Details

The “point” of these functions is that if A and C (or B and C) are identical nonpositive integers, a warning needs to be given because the function is defined as the appropriate limit and one needs to be sure that both A and C approach that limit at the same speed.

Function hypergeo_AorB_nonpos_int() is a convenience wrapper for hypergeo_A_nonpos_int().

Author(s)

Robin K. S. Hankin

References

M. Abramowitz and I. A. Stegun 1965. Handbook of mathematical functions. New York: Dover

See Also

hypergeo

Examples

jjR1 <- hypergeo(-4, pi, 2.2 , 1+6i)
jjR2 <- hypergeo(pi, -4, 2.2 , 1+6i)  # former bug
jjM <- 3464.1890402837334002-353.94143580568566281i  # value given by Mathematica 

[Package hypergeo version 1.2-14 Index]