hypergeo_cover1 {hypergeo}	R Documentation

Hypergeometric functions for special values of the parameters

Description

Hypergeometric functions for special values of the parameters

Usage

hypergeo_cover1(A, B, m, z, tol = 0, maxiter = 2000, method = "a", give = FALSE)
hypergeo_cover2(A, C, m, z, tol = 0, maxiter = 2000, method = "a", give = FALSE)
hypergeo_cover3(A, n, m, z, tol = 0, maxiter = 2000, method = "a", give = FALSE)

Arguments

A,B,C
parameters for the hypergeometric function

m,n
Integers (positive or negative)

z
Primary complex argument

tol,maxiter
Numerical arguments passed to genhypergeo()

method
Method, passed to f15.3.10() (qv)

give
Boolean with TRUE meaning to return the choice of helper function used (eg f15.3.7()), and default FALSE meaning to return the hypergeometric function's value

Details

These functions deal with the exceptional cases listed on page 559-560.

Function hypergeo_cover1() deals with the case C=A+B +/- m, m=0,1,2,...

Function hypergeo_cover2() deals with the case B=A +/- m, m=0,1,2,...

Function hypergeo_cover3() deals with the case C-A=0,-1,-2,... [elementary] and C-A=1,2,... [not covered by AMS-55]

Note

Function hypergeo_cover3() is required because the “limiting process” mentioned on p560, just after 15.3.14, is not explicit. Which is why it dispatches to w07.23.06.0026.01() and w07.23.06.0031.01(), documented at wolfram.

Author(s)

Robin K. S. Hankin

References

M. Abramowitz and I. A. Stegun 1965. Handbook of mathematical functions. New York: Dover

See Also

hypergeo,f15.3.10,wolfram

Examples

# Test hypergeo_cover1():
jjR <- hypergeo(pi,pi/2,3*pi/2-4, z=0.1+0.2i)
jjM <- 0.53745229690249593045 + 1.8917456473240515664i

# Test hypergeo_cover2():
jjM <- -0.15888831928748121465e-5 + 0.40339599711492215912e-4i
jjR <- hypergeo(pi,pi+2, 1.1 , 1+10i)  # This is 15.3.13
stopifnot(Mod(jjR-jjM)<1e-10)

# Test hypergeo_cover3()
jjM <- -0.24397135980533720308e-1 + 0.28819643319432922231i
jjR <- hypergeo(pi, 1.4, pi+4, 1+6i)
stopifnot(Mod(jjR-jjM)<1e-10)

[Package hypergeo version 1.2-14 Index]