hypergeo_powerseries {hypergeo}	R Documentation

The hypergeometric function as determined by power series

Description

The hypergeometric function as determined by infinite (hypergeo_powerseries()) or finite (hypergeo_taylor()) power series

Usage

hypergeo_powerseries(A, B, C, z, tol = 0, maxiter = 2000)

Arguments

A,B,C
Parameters of the hypergeometric function

z
Primary complex argument

tol,maxiter
Numerical arguments

Details

Function hypergeo_powerseries() is the primary decision-making function of the package. It is this function that detects degenerate cases of the three parameters and dispatches accordingly. Non-degenerate cases are sent to function hypergeo_general().

Function hypergeo_taylor() deals with cases where the hypergeometric function is a polynomial.

Author(s)

Robin K. S. Hankin

References

M. Abramowitz and I. A. Stegun 1965. Handbook of mathematical functions. New York: Dover

See Also

hypergeo,genhypergeo

Examples

jjR <- hypergeo(pi,-4,2.2,1+5i)
jjM <- 1670.8287595795885335 - 204.81995157365381258i

[Package hypergeo version 1.2-14 Index]