i15.3.6 {hypergeo}	R Documentation

Helper functions

Description

Helper functions for equations 15.3.6-15.3.9

Usage

i15.3.6(A, B, C)
i15.3.7(A, B, C)
i15.3.8(A, B, C)
i15.3.9(A, B, C)
j15.3.6(A, B, C)
j15.3.7(A, B, C)
j15.3.8(A, B, C)
j15.3.9(A, B, C)

Arguments

A,B,C
Parameters of the hypergeometric function

Details

These functions are named for their equation numbers in Abramowitz and Stegun.

Functions i15.3.?() return the factors at the beginning of equations 15.3.6-9. These functions return zero if the denominator is infinite (because it includes a gamma function of a nonpositive integer).

Functions j15.3.?() check for the appropriate arguments of the gamma function being nonpositive integers.

Author(s)

Robin K. S. Hankin

References

M. Abramowitz and I. A. Stegun 1965. Handbook of mathematical functions. New York: Dover

See Also

hypergeo

Examples

i15.3.6(1.1, 3.2, pi)

[Package hypergeo version 1.2-14 Index]