is.nonpos {hypergeo}	R Documentation

Various utilities

Description

Various utilities needing nonce functions

Usage

is.near_integer(i, tol=getOption("tolerance"))
is.nonpos(i)
is.zero(i)
isgood(x, tol)
thingfun(z, complex=FALSE)
crit(...)
lpham(x,n)

Arguments

i
Numerical vector of suspected integers

tol
Tolerance

x
Argument to isgood() and lpham()

z
Complex vector

complex
In function thingfun(), Boolean with default FALSE meaning to return the modulus of the transforms and TRUE meaning to return the complex values themselves

n
second argument to lpham()

...
Ignored

Details

Function is.near_integer(i) returns TRUE if i is “near” [that is, within tol] an integer; if the option is unset then 1e-11 is used.

Function is.nonpos() returns TRUE if i is near a nonpositive integer

Function is.zero() returns TRUE if i is, er, near zero

Function isgood() checks for all elements of x having absolute values less than tol

Function thingfun() transforms input vector z by each of the six members of the anharmonic group, viewed as a subgroup of the Mobius group of functions. It returns a real six-column matrix with columns being the modulus of z,z/(z-1),1-z,1/z,1/(1-z),1-1/z. These six columns correspond to the primary argument in equations 15.3.3 to 15.3.9, p551 of AMS-55

Function crit() returns the two critical points, 1/2 +/- sqrt(3)i/2. These points have unit modulus as do their six transforms by thingfun()

Function lpham() returns the log of the Pochhammer function log(Gamma(x+n)/Gamma(x))

Note

Function isgood() uses zero as the default tolerance (argument tol passed in from hypergeo()); compare the different meaning of tol used in is.near_integer().

Here, “integer” means one of the sequence 0, +/-1, +/-2, ... [ie not the Gaussian integers].

Author(s)

Robin K. S. Hankin

Examples

is.near_integer(-3)

is.zero(4)

[Package hypergeo version 1.2-14 Index]