residue {hypergeo}	R Documentation

Evaluation of the hypergeometric function using the residue theorem

Description

Expansion of the hypergeometric function using the residue theorem; useful for when the primary argument is close to the critical points 0.5+/-i.sqrt(3)/2

Usage

hypergeo_residue_general(A, B, C, z, r, O=z, tol=0, maxiter=2000)
hypergeo_residue_close_to_crit_single(A, B, C, z, strategy='A', tol=0, maxiter=2000)
hypergeo_residue_close_to_crit_multiple(A, B, C, z, strategy='A', tol=0, maxiter=2000)

Arguments

A,B,C
Parameters (real or complex)

z
Complex argument

tol,maxiter
tolerance and maximum number of iterations (passed to hypergeo())

r,O
Radius and center of circle to integrate over

strategy
Indicates which strategy to use. Strategy ‘A’ means to use the critical point as the centre of the circle and strategy ‘B’ means to use z

Details

These functions are not really intended for the user; hypergeo() uses hypergeo_residue_close_to_crit_multiple() when abs(z-c) is less than 0.1 (hardwired) for c being either of the two critical points. Infinite regress is avoided because the contour is always more than this distance from the critical points.

These functions use the residue theorem f(z_0)=int f(z)dz/(z-z_0) to evaluate the hypergeometric function near the two critical points 0.5+/-i.sqrt(3)/2. These points are problematic because all of the transformations listed under thingfun() take the points either to themselves or each other.

At these points the ratio of successive terms in the hypergeometric series tends to one and thus numerical summation is difficult.

The hypergeometric function, however, is not at all badly behaved near these critical points (see examples); but OTOH there do not seem to be any identities for the hypergeometric function at these points.

I have not investigated in detail whether strategy ‘A’ or ‘B’ is better. I would expect that ‘A’ is faster but ‘B’ more accurate, on the grounds that ‘A’ uses a contour whose closest approach to the critical point is further than that of ‘B’; but ‘B’ uses a contour which does not vary in distance from z.

But both seem to be fairly accurate and fairly fast, and I have not systematically investigated the pros and cons.

Note

The residue theorem appears to be absurdly accurate for numerical evaluation

Author(s)

Robin K. S. Hankin

References

W. Buhring 1987. “An analytic continuation of the hypergeometric series”, Siam J. Math. Anal. 18(3)

See Also

buhring

Examples

c1 <- 1/2-sqrt(3)/2i
c2 <- 1/2+sqrt(3)/2i

a1_R <- hypergeo(1/2,1/3,pi,c1)
a1_M <- 1.0154051314906669 + 0.0544835896509068i

x <- y <- seq(from=-0.1,to=0.1,len=100)
elliptic::view(x,y,hypergeo(1/2,1,1/3,outer(x,1i*y,"+")))

[Package hypergeo version 1.2-14 Index]