shanks {hypergeo}	R Documentation

Evaluation of the hypergeometric function using Shanks's method

Description

Evaluation of the hypergeometric function using Shanks transformation of successive sums

Usage

hypergeo_shanks(A,B,C,z,maxiter=20)
genhypergeo_shanks(U,L,z,maxiter=20)
shanks(Last,This,Next)

Arguments

A,B,C
Parameters (real or complex)

U,L
Upper and lower vectors

z
Primary complex argument

maxiter
Maximum number of iterations

Last,This,Next
Three successive convergents

Details

The Shanks transformation of successive partial sums is

S(n)=\frac{A_{n+1}A_{n-1}-A_n^2}{A_{n+1}-2A_n+A_{n-1}}

and if the A_n tend to a limit then the sequence S(n) often converges more rapidly than A_n. However, the denominator is susceptible to catastrophic rounding under fixed-precision arithmetic and it is difficult to know when to stop iterating.

Note

The

Author(s)

Robin K. S. Hankin

References

Shanks, D. (1955). “Non-linear transformation of divergent and slowly convergent sequences”, Journal of Mathematics and Physics 34:1-42

See Also

buhring

Examples

hypergeo_shanks(1/2,1/3,pi,z= 0.1+0.2i)

[Package hypergeo version 1.2-14 Index]